Created routes manifest; rename 'notifications' to 'native notifications'; rename 'errors' to 'notifications'
