+++ src/components/Notification/Notification.js
@@ -2,28 +2,65 @@
  * @module components/Notification
  */
 
+import React, { Component } from 'react';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { withTranslation } from 'react-i18next';
 
-import socketIOEventHandlerMapping from '../../socket-io/index'
-import socket from '../../socket-io';
+import { errorActions } from '../../actions';
 
-function Notification(props) {
-  const { dispatch } = props;
-  socketIOEventHandlerMapping.forEach((handler, event) => {
-    socket.on(event, data => handler(dispatch, data));
-  });
-  return null;
+class Notification extends Component {
+
+  constructor(props) {
+    super(props);
+    this.handleClose = this.handleClose.bind(this);
+    this.handleExited = this.handleExited.bind(this);
+  }
+
+  handleClose(event, reason) {
+    if (reason === 'clickaway') {
+      return;
+    }
+    this.props.dispatch(errorActions.closed());
+  }
+
+  handleExited() {
+    this.props.dispatch(errorActions.dequeued());
+  }
+
+  render() {
+    const { errors, t } = this.props;
+
+    return (
+      <div>
+        {
+          errors.all.map((error, index) => {
+            const variant = error.variant || 'danger';
+            return (
+              <div style={{width: '400px', position: 'absolute', top: 10, right: 10}} dismissible key={index} variant={variant}>{t(error.translationKey)}</div>
+            );
+          })
+        }
+      </div>
+    );
+  }
+
 }
 
 Notification.propTypes = {
-  dispatch: PropTypes.func.isRequired
+  dispatch: PropTypes.func.isRequired,
+  errors: PropTypes.object.isRequired
+}
+
+function mapStateToProps(state) {
+  return {
+    errors: state.errors
+  }
 }
 
 export default compose(
   // withStyles(styles),
   withTranslation(),
-  connect()
+  connect(mapStateToProps)
 )(Notification);
