--- src/types/error.js
@@ -1,9 +1,0 @@
-/**
- * @module types/error
- */
-
-export default {
-  ERROR_ENQUEUED: 'ERROR_ENQUEUED',
-  ERROR_DEQUEUED: 'ERROR_DEQUEUED',
-  ERROR_CLOSED: 'ERROR_CLOSED'
-}
