+++ src/types/index.js
@@ -7,35 +7,35 @@
 /**********************************/
 
 import _authenticationTypes from './authentication';
-import _errorTypes from './error';
+import _notificationTypes from './notification';
 import _googleTypes from './google';
 // import _eventTypes from './event';
 import _i18nTypes from './i18n';
 import _meTypes from './me';
 import _microsoftTypes from './microsoft';
-import _notificationTypes from './notification';
+import _nativeNotificationTypes from './native-notification';
 import _sdkTypes from './sdk';
 import _userTypes from './user';
 
 export default {
   authentication: _authenticationTypes,
-  error: _errorTypes,
+  notification: _notificationTypes,
   // event: _eventTypes,
   i18n: _i18nTypes,
   me: _meTypes,
   microsoft: _microsoftTypes,
-  notification: _notificationTypes,
+  nativeNotification: _nativeNotificationTypes,
   sdk: _sdkTypes,
   user: _userTypes
 };
 
 export const authenticationTypes = _authenticationTypes;
-export const errorTypes = _errorTypes;
+export const notificationTypes = _notificationTypes;
 export const googleTypes = _googleTypes;
 // export const eventTypes = _eventTypes;
 export const i18nTypes = _i18nTypes;
 export const meTypes = _meTypes;
 export const microsoftTypes = _microsoftTypes;
-export const notificationTypes = _notificationTypes;
+export const nativeNotificationTypes = _nativeNotificationTypes;
 export const sdkTypes = _sdkTypes;
 export const userTypes = _userTypes;
