+++ src/types/native-notification.js
@@ -1,0 +1,12 @@
+/**
+ * @module types/native-notification
+ */
+
+export default {
+  NATIVE_NOTIFICATION_PERMISSION_REQUESTED: 'NATIVE_NOTIFICATION_PERMISSION_REQUESTED',
+  NATIVE_NOTIFICATION_PERMISSION_GRANTED: 'NATIVE_NOTIFICATION_PERMISSION_GRANTED',
+  NATIVE_NOTIFICATION_PERMISSION_DENIED: 'NATIVE_NOTIFICATION_PERMISSION_DENIED',
+  NATIVE_NOTIFICATION_STARTED: 'NATIVE_NOTIFICATION_STARTED',
+  NATIVE_NOTIFICATION_FAILED: 'NATIVE_NOTIFICATION_FAILED',
+  NATIVE_NOTIFICATION_SUCCEEDED: 'NATIVE_NOTIFICATION_SUCCEEDED'
+}
