+++ src/types/notification.js
@@ -3,10 +3,7 @@
  */
 
 export default {
-  NOTIFICATION_PERMISSION_REQUESTED: 'NOTIFICATION_PERMISSION_REQUESTED',
-  NOTIFICATION_PERMISSION_GRANTED: 'NOTIFICATION_PERMISSION_GRANTED',
-  NOTIFICATION_PERMISSION_DENIED: 'NOTIFICATION_PERMISSION_DENIED',
-  NOTIFICATION_STARTED: 'NOTIFICATION_STARTED',
-  NOTIFICATION_FAILED: 'NOTIFICATION_FAILED',
-  NOTIFICATION_SUCCEEDED: 'NOTIFICATION_SUCCEEDED'
+  NOTIFICATION_ENQUEUED: 'NOTIFICATION_ENQUEUED',
+  NOTIFICATION_DEQUEUED: 'NOTIFICATION_DEQUEUED',
+  NOTIFICATION_CLOSED: 'NOTIFICATION_CLOSED'
 }
